requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
sqlmodel>=0.0.16
pydantic>=2.7.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
pydantic-settings>=2.0.0
chromadb>=0.4.0
pypdf>=3.0.0
python-docx>=1.0.0
tiktoken>=0.5.0
langchain-community>=0.0.10
sentence-transformers>=2.2.0
minio>=7.2.0
